+++ share/ruby-install/ruby-install.sh
@@ -2,210 +2,25 @@
 
 shopt -s extglob
 
-ruby_install_version="0.5.1"
+ruby_install_version="0.6.0"
 ruby_install_dir="${BASH_SOURCE[0]%/*}"
-
-source "$ruby_install_dir/versions.sh"
+ruby_install_cache_dir="${XDG_CACHE_HOME:-$HOME/.cache}/ruby-install"
 
 rubies=(ruby jruby rbx maglev mruby)
 patches=()
 configure_opts=()
 make_opts=()
 
-#
-# Auto-detect the package manager.
-#
-if   command -v pacman  >/dev/null; then package_manager="pacman"
-elif command -v dnf     >/dev/null; then package_manager="dnf"
-elif command -v yum     >/dev/null; then package_manager="yum"
-elif command -v apt-get >/dev/null; then package_manager="apt"
-elif command -v port    >/dev/null; then package_manager="port"
-elif command -v brew    >/dev/null; then package_manager="brew"
+if (( UID == 0 )); then
+	src_dir="/usr/local/src"
+	rubies_dir="/opt/rubies"
+else
+	src_dir="$HOME/src"
+	rubies_dir="$HOME/.rubies"
 fi
 
-#
-# Auto-detect the downloader.
-#
-if   command -v wget >/dev/null; then downloader="wget"
-elif command -v curl >/dev/null; then downloader="curl"
-fi
-
-#
-# Only use sudo if already root.
-#
-if (( UID == 0 )); then sudo=""
-else                    sudo="sudo"
-fi
-
-#
-# Prints a log message.
-#
-function log()
-{
-	if [[ -t 1 ]]; then
-		echo -e "\x1b[1m\x1b[32m>>>\x1b[0m \x1b[1m$1\x1b[0m"
-	else
-		echo ">>> $1"
-	fi
-}
-
-#
-# Prints a warn message.
-#
-function warn()
-{
-	if [[ -t 1 ]]; then
-		echo -e "\x1b[1m\x1b[33m***\x1b[0m \x1b[1m$1\x1b[0m" >&2
-	else
-		echo "*** $1" >&2
-	fi
-}
-
-#
-# Prints an error message.
-#
-function error()
-{
-	if [[ -t 1 ]]; then
-		echo -e "\x1b[1m\x1b[31m!!!\x1b[0m \x1b[1m$1\x1b[0m" >&2
-	else
-		echo "!!! $1" >&2
-	fi
-}
-
-#
-# Prints an error message and exists with -1.
-#
-function fail()
-{
-	error "$@"
-	exit -1
-}
-
-#
-# Searches a file for a key and echos the value.
-# If the key cannot be found, the third argument will be echoed.
-#
-function fetch()
-{
-	local file="$ruby_install_dir/$1.txt"
-	local key="$2"
-	local line
-
-	while IFS="" read -r line; do
-		if [[ "$line" == "$key:"* ]]; then
-			echo "${line##$key:*([[:space:]])}"
-		fi
-	done < "$file"
-}
-
-function install_packages()
-{
-	case "$package_manager" in
-		apt)	$sudo apt-get install -y "$@" || return $? ;;
-		dnf)	$sudo dnf install -y "$@" || return $?     ;;
-		yum)	$sudo yum install -y "$@" || return $?     ;;
-		port)   $sudo port install "$@" || return $?       ;;
-		brew)
-			local brew_owner="$(/usr/bin/stat -f %Su "$(command -v brew)")"
-			sudo -u "$brew_owner" brew install "$@" ||
-			sudo -u "$brew_owner" brew upgrade "$@" || return $?
-			;;
-		pacman)
-			local missing_pkgs=($(pacman -T "$@"))
-
-			if (( ${#missing_pkgs[@]} > 0 )); then
-				$sudo pacman -S "${missing_pkgs[@]}" || return $?
-			fi
-			;;
-		"")	warn "Could not determine Package Manager. Proceeding anyway." ;;
-	esac
-}
-
-#
-# Downloads a URL.
-#
-function download()
-{
-	local url="$1"
-	local dest="$2"
-
-	[[ -d "$dest" ]] && dest="$dest/${url##*/}"
-	[[ -f "$dest" ]] && return
-
-	case "$downloader" in
-		wget) wget -c -O "$dest.part" "$url" || return $?         ;;
-		curl) curl -f -L -C - -o "$dest.part" "$url" || return $? ;;
-		"")
-			error "Could not find wget or curl"
-			return 1
-			;;
-	esac
-
-	mv "$dest.part" "$dest" || return $?
-}
-
-#
-# Extracts an archive.
-#
-function extract()
-{
-	local archive="$1"
-	local dest="${2:-${archive%/*}}"
-
-	case "$archive" in
-		*.tgz|*.tar.gz) tar -xzf "$archive" -C "$dest" || return $? ;;
-		*.tbz|*.tbz2|*.tar.bz2)	tar -xjf "$archive" -C "$dest" || return $? ;;
-		*.zip) unzip "$archive" -d "$dest" || return $? ;;
-		*)
-			error "Unknown archive format: $archive"
-			return 1
-			;;
-	esac
-}
-
-#
-# Loads function.sh for the given Ruby.
-#
-function load_ruby()
-{
-	ruby_dir="$ruby_install_dir/$ruby"
-
-	if [[ ! -d "$ruby_dir" ]]; then
-		echo "ruby-install: unsupported ruby: $ruby" >&2
-		return 1
-	fi
-
-	local absolute_version="$(
-	  resolve_version "$ruby_version" \
-		          "$ruby_dir/versions.txt" \
-			  "$ruby_dir/stable.txt"
-	)"
-
-	if [[ -n "$absolute_version" ]]; then
-		ruby_version="$absolute_version"
-	else
-		warn "Unknown $ruby version: $ruby_version"
-	fi
-
-	source "$ruby_install_dir/functions.sh" || return $?
-	source "$ruby_dir/functions.sh" || return $?
-}
-
-#
-# Prints Rubies supported by ruby-install.
-#
-function known_rubies()
-{
-	local ruby
-
-	echo "Latest ruby versions:"
-
-	for ruby in "${rubies[@]}"; do
-		echo "  $ruby:"
-		sed -e 's/^/    /' "$ruby_install_dir/$ruby/stable.txt" || return $?
-	done
-}
+source "$ruby_install_dir/util.sh"
+source "$ruby_install_dir/ruby-versions.sh"
 
 #
 # Prints usage information for ruby-install.
@@ -236,6 +51,7 @@ Options:
 	--no-extract		Do not re-extract the downloaded Ruby archive
 	--no-install-deps	Do not install build dependencies before installing Ruby
 	--no-reinstall  	Skip installation if another Ruby is detected in same location
+	-L, --latest		Downloads the latest ruby versions and checksums
 	-V, --version		Prints the version
 	-h, --help		Prints this message
 
@@ -250,6 +66,30 @@ Examples:
 	$ ruby-install -p https://raw.github.com/gist/4136373/falcon-gc.diff ruby 1.9.3
 
 USAGE
+}
+
+#
+# Parses a "ruby-version" string.
+#
+function parse_ruby()
+{
+	local arg="$1"
+
+	case "$arg" in
+		*-*)
+			ruby="${arg%%-*}"
+			ruby_version="${arg#*-}"
+			;;
+		*)
+			ruby="${arg}"
+			ruby_version=""
+			;;
+	esac
+
+	if [[ ! "${rubies[@]}" == *"$ruby"* ]]; then
+		error "Unknown ruby: $ruby"
+		return 1
+	fi
 }
 
 #
@@ -335,6 +175,10 @@ function parse_options()
 				no_reinstall=1
 				shift
 				;;
+			-L|--latest)
+				force_update=1
+				shift
+				;;
 			-V|--version)
 				echo "ruby-install: $ruby_install_version"
 				exit
@@ -360,23 +204,60 @@ function parse_options()
 	done
 
 	case ${#argv[*]} in
-		2)
-			ruby="${argv[0]}"
-			ruby_version="${argv[1]}"
-			;;
-		1)
-			ruby="${argv[0]}"
-			ruby_version=""
-			;;
-		0)
-			echo "ruby-install: too few arguments" >&2
-			usage 1>&2
-			return 1
-			;;
+		2)	parse_ruby "${argv[0]}-${argv[1]}" || return $? ;;
+		1)	parse_ruby "${argv[0]}" || return $? ;;
+		0)	return 0 ;;
 		*)
 			echo "ruby-install: too many arguments: ${argv[*]}" >&2
 			usage 1>&2
 			return 1
 			;;
 	esac
+}
+
+#
+# Prints Rubies supported by ruby-install.
+#
+function list_rubies()
+{
+	local ruby
+
+	for ruby in "${rubies[@]}"; do
+		if [[ $force_update -eq 1 ]] ||
+		   are_ruby_versions_missing "$ruby"; then
+			log "Downloading latest $ruby versions ..."
+			download_ruby_versions "$ruby"
+		fi
+	done
+
+	echo "Stable ruby versions:"
+	for ruby in "${rubies[@]}"; do
+		echo "  $ruby:"
+		stable_ruby_versions "$ruby" | sed -e 's/^/    /' || return $?
+	done
+}
+
+#
+# Initializes variables.
+#
+function init()
+{
+	local fully_qualified_version="$(lookup_ruby_version "$ruby" "$ruby_version")"
+
+	if [[ -n "$fully_qualified_version" ]]; then
+		ruby_version="$fully_qualified_version"
+	else
+		warn "Unknown $ruby version $ruby_version. Proceeding anyways ..."
+	fi
+
+	ruby_cache_dir="$ruby_install_cache_dir/$ruby"
+	install_dir="${install_dir:-$rubies_dir/$ruby-$ruby_version}"
+
+	source "$ruby_install_dir/functions.sh"       || return $?
+	source "$ruby_install_dir/$ruby/functions.sh" || return $?
+
+	ruby_md5="${ruby_md5:-$(ruby_checksum_for "$ruby" md5 "$ruby_archive")}"
+	ruby_sha1="${ruby_sha1:-$(ruby_checksum_for "$ruby" sha1 "$ruby_archive")}"
+	ruby_sha256="${ruby_sha256:-$(ruby_checksum_for "$ruby" sha256 "$ruby_archive")}"
+	ruby_sha512="${ruby_sha512:-$(ruby_checksum_for "$ruby" sha512 "$ruby_archive")}"
 }