+++ share/ruby-install/versions.sh
@@ -5,6 +5,10 @@ function is_known_version()
 	local file="$1"
 	local version="$2"
 
+	if [[ ! -f "$file" ]] || [[ -z "$version" ]]; then
+		return 1
+	fi
+
 	grep -q -x "$version" "$file"
 }
 
@@ -12,6 +16,10 @@ function latest_version()
 {
 	local file="$1"
 	local key="$2"
+
+	if [[ ! -f "$file" ]]; then
+		return 1
+	fi
 
 	if [[ -z "$key" ]]; then
 		tail -n 1 "$file"
@@ -30,21 +38,5 @@ function latest_version()
 		echo -n "$match"
 	else
 		return 1
-	fi
-}
-
-function resolve_version()
-{
-	local version="$1"
-	local versions_file="$2"
-	local latest_versions_file="$3"
-
-	if [[ -z "$version" ]]; then
-		latest_version "$latest_versions_file"
-	elif is_known_version "$versions_file" "$version"; then
-		echo -n "$version"
-	else
-		latest_version "$latest_versions_file" "$version" ||
-		echo -n "$version"
 	fi
 }