+++ test/functions_tests/apply_patches_test.sh
@@ -1,13 +1,13 @@
 #!/usr/bin/env bash
 
 . ./test/helper.sh
+. ./share/ruby-install/functions.sh
+
+ruby_src_dir="ruby-1.9.3-p448"
+patches=("$src_dir/$ruby_src_dir/falcon-gc.diff")
 
 function setUp()
 {
-	src_dir="$PWD/test/src"
-	ruby_src_dir="ruby-1.9.3-p448"
-	patches=("$src_dir/$ruby_src_dir/falcon-gc.diff")
-
 	mkdir -p "$src_dir/$ruby_src_dir"
 	echo "diff -Naur $ruby_src_dir.orig/test $ruby_src_dir/test
 --- $ruby_src_dir.orig/test 1970-01-01 01:00:00.000000000 +0100