+++ test/ruby-install_tests/parse_options_test.sh
@@ -16,9 +16,11 @@ function setUp()
 
 function test_parse_options_with_no_arguments()
 {
-	parse_options >/dev/null 2>&1
+	parse_options
 
-	assertEquals "did not return 1" 1 $?
+	assertEquals "did not return 0" 0 $?
+	assertNull "did not leave \$ruby blank" "$ruby"
+	assertNull "did not leave \$ruby_version blank" "$ruby_version"
 }
 
 function test_parse_options_with_invalid_options()