+++ test/versions_test.sh
@@ -25,6 +25,24 @@ function test_is_known_version()
 		     0 $?
 }
 
+function test_is_known_version_with_invalid_file()
+{
+	local version="2.0.0-p576"
+
+	is_known_version "./test/foo/bar" "$version"
+
+	assertEquals "did not return an error" 1 $?
+}
+
+function test_is_known_version_with_empty_version()
+{
+	local version=""
+
+	is_known_version "$versions_file" "$version"
+
+	assertEquals "did not return an error" 1 $?
+}
+
 function test_is_known_version_with_invalid_version()
 {
 	local version="1.2.3"
@@ -60,44 +78,6 @@ function test_latest_version_with_unknown_version()
 	latest_version "$stable_file" "$unknown_version"
 
 	assertEquals "did not return an error" 1 $?
-}
-
-function test_resolve_version_with_exact_version()
-{
-	local version="2.0.0-p0"
-
-	assertEquals "did not return the exact version" \
-		     "$version" \
-		     "$(resolve_version "$version" "$versions_file" "$stable_file")"
-}
-
-function test_resolve_version_with_short_version()
-{
-	local version="2.0"
-	local expected_version="2.0.0-p481"
-
-	assertEquals "did not return the exact version" \
-		     "$expected_version" \
-		     "$(resolve_version "$version" "$versions_file" "$stable_file")"
-}
-
-function test_resolve_version_with_empty_string()
-{
-	local version=""
-	local expected_version="2.1.3"
-
-	assertEquals "did not return the exact version" \
-		     "$expected_version" \
-		     "$(resolve_version "$version" "$versions_file" "$stable_file")"
-}
-
-function test_resolve_version_with_new_version()
-{
-	local new_version="3.0.0"
-
-	assertEquals "did not return the new version" \
-		     "$new_version" \
-		     "$(resolve_version "$new_version" "$versions_file" "$stable_file")"
 }
 
 function oneTimeTearDown()